Refractor conditions to its own file
